Use third party library to determine chrome flags
The library go-rod has a well tested set of options which
are translated into parameters for the chrome library.

It also uses all common paths on different OSes to find
an installed chrome and can download one on the fly if
none is found.

This combination allows for a much more stable chrome
process which can fix a whole range of issues that exist
with the current approach.

Relates: #22, #23